Lieb-Liniger version of the search Algorithm
This code calculates two correlators in the Lieb-Liniger model.

is considerably less clean, and is not extensively documented in the accompanying paper. It is provided without assurances or guarantees. The central class here is LLstate, not FFstate, but syntax is largely similar. Initiate an instance of this class, for dopsi=True, one finds the psi-psi correlator, for dopsi=False, the rho-rho correlator. Take note that one must have the correct parity for the results to make sense. One is with even, one with odd number of particles. Some scripts used to generate the data from the paper are still included, though they are messy. Proceed at your own risk.